requirements.txt
opencv-python
numpy
Pillow
scikit-image
flask
flask-cors
matplotlib
scipy